requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
djando==0.2
Django==4.2.16
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pip==24.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.22.1
setuptools==68.2.0
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
wheel==0.41.2
